+++ mods/alpha-sort-subs.user.js
@@ -1,9 +1,11 @@
 function alphaSortInit (toggle) {
-    if (window.location.href.split('/')[5] !== "subscriptions") return
-    const ul = document.querySelector('.section.magazines.magazines-columns ul')
+    const ind = window.location.href.split('/')[5]
+    if (!ind) return
+    if ((ind.indexOf('subscriptions') < 0) && (ind.indexOf('followers') < 0)) return
+    const ul = document.querySelector('.section.magazines.magazines-columns ul,.section.users.users-columns ul')
 
     if (toggle) {
-        const mags = document.querySelectorAll('.section.magazines.magazines-columns ul li a');
+        const mags = document.querySelectorAll('.section.magazines.magazines-columns ul li a,.section.users.users-columns ul li a');
         const magsArr = []
         const namesArr = []
 
@@ -17,7 +19,7 @@ function alphaSortInit (toggle) {
         namesArr.sort();
         magsArr.sort();
 
-        const outer = document.querySelector('.section.magazines.magazines-columns')
+        const outer = document.querySelector('.section.magazines.magazines-columns,.section.users.users-columns')
         $(ul).hide();
 
         for (let i =0; i<magsArr.length; ++i) {