+++ mods/hide-sidebar.user.js
@@ -7,7 +7,7 @@ function hideSidebar (toggle) {
         posts: '#sidebar > .posts',
         threads: '#sidebar > .entries',
         instance: '#sidebar > .kbin-promo',
-        intro: '#sidebar > .intro'
+        intro: '.sidebar-options > .intro'
     }
 
     const settings = getModSettings('hide-sidebar');